backend/regulatory_change_agent/requirements.txt
# openai
# pdfplumber
# python-docx
# fastapi
# uvicorn
# python-multipart

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.40.6
python-dotenv==1.0.1